Data base y data clean

+++ src/ProyectoTesisModels/Modelos/IdClean.cs
@@ -1,5 +1,6 @@
 ﻿using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
+using System;
 
 namespace ProyectoTesisModels.Modelos
 {
@@ -7,6 +8,6 @@ namespace ProyectoTesisModels.Modelos
     public class IdClean
     {
         public ObjectId id { get; set; }
-        public int user { get; set; }
+        public Int64 user { get; set; }
     }
 }
